Add role localization to Team members display
Previously, roles were simply a static string in English.
Now, the roles are stored as a list of keys that correspond to localized
strings. These strings are managed in each language's respective JSON file.
This change allows for the website to be displayed in numerous languages
without losing the context of each team member's role. It also makes adding
new roles or modifying existing ones much simpler.